Change package name from `com.jakipatryk.spark.persistenthomology` to `io.github.jakipatryk.sparkpersistenthomology`
