--- src/main/scala/com/jakipatryk/spark/persistenthomology/distances/DistanceCalculator.scala
@@ -1,7 +1,0 @@
-package com.jakipatryk.spark.persistenthomology.distances
-
-trait DistanceCalculator extends Serializable {
-
-  def calculateDistance(v1: Vector[Double], v2: Vector[Double]): Double
-
-}
