--- src/main/scala/com/jakipatryk/spark/persistenthomology/distances/EuclideanDistanceCalculator.scala
@@ -1,12 +1,0 @@
-package com.jakipatryk.spark.persistenthomology.distances
-
-object EuclideanDistanceCalculator extends DistanceCalculator {
-
-  override def calculateDistance(v1: Vector[Double], v2: Vector[Double]): Double = math.sqrt(
-      v1
-        .zip(v2)
-        .map { case (a, b) => (a - b) * (a - b) }
-        .sum
-  )
-
-}
