--- src/main/scala/com/jakipatryk/spark/persistenthomology/matrixreduction/package.scala
@@ -1,9 +1,0 @@
-package com.jakipatryk.spark.persistenthomology
-
-import org.apache.spark.rdd.RDD
-
-package object matrixreduction {
-
-  case class BoundaryMatrix(rdd: RDD[(Key, Chain)]) extends AnyVal
-
-}
