+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/distances/DistanceCalculator.scala
@@ -1,0 +1,7 @@
+package io.github.jakipatryk.sparkpersistenthomology.distances
+
+trait DistanceCalculator extends Serializable {
+
+  def calculateDistance(v1: Vector[Double], v2: Vector[Double]): Double
+
+}
