+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/matrixreduction/package.scala
@@ -1,0 +1,9 @@
+package io.github.jakipatryk.sparkpersistenthomology
+
+import org.apache.spark.rdd.RDD
+
+package object matrixreduction {
+
+  case class BoundaryMatrix(rdd: RDD[(Key, Chain)]) extends AnyVal
+
+}
